requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
pytz==2022.1
requests==2.28.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
